🚧 Fix sign in logic + MSW mocking + change structure
