+++ .env
@@ -1,3 +1,3 @@
 NODE_ENV=development
-NEXT_PUBLIC_SERVER_API_URL=https://your-server-url.com/api
+NEXT_PUBLIC_SERVER_API_URL='http://localhost:3000'
 NEXT_PUBLIC_API_TOKEN=your_api_token
