+++ src/app/signUp/steps/step2Phone.tsx
