+++ src/app/signUp/steps/step7Password.tsx
