+++ src/mocks/handler.js
@@ -11,8 +11,8 @@ import { paginatedTeamMenus, teamMenus } from './mockData/teamMenus';
 
 export const handler = [
   // Handler for restaurant list
-  http.get(RESTAURANT_LIST_API_URL, (req) => {
-    const { request } = req;
+  http.get(RESTAURANT_LIST_API_URL, async (req) => {
+    const { request } = await req;
     const urlString = request.url.toString();
 
     try {
@@ -185,13 +185,68 @@ export const handler = [
     const meetingId = Number(req.params.meetingId);
     const purchaseId = Number(req.params.purchaseId);
 
-    const teamMenu = teamMenus.filter((teamMenu) => teamMenu.meetingId === meetingId);
+    const teamMenu = teamMenus.filter(
+      (teamMenu) => teamMenu.meetingId === meetingId,
+    );
     const teamPurchase = teamMenu.find((p) => p.purchaseId === purchaseId);
 
     if (teamPurchase) {
       return HttpResponse.json(teamPurchase);
     }
 
-    return HttpResponse.status(404).json({ message: 'Team Menu Item not found' });
+    return HttpResponse.status(404).json({
+      message: 'Team Menu Item not found',
+    });
+  }),
+
+  // Handler for user sign in
+  http.post('/api/users/signin', async ({ request }) => {
+    try {
+      const { email, password } = await request.json();
+
+      if (email === 'test@example.com' && password === 'password123') {
+        const headers = new Headers();
+        headers.set('Authorization', 'mocked-jwt-token');
+        headers.set('Refresh', 'mocked-refresh-token');
+
+        return HttpResponse.json(
+          { message: '로그인 성공' },
+          { status: 200, headers },
+        );
+      }
+
+      return HttpResponse.json(
+        { message: '이메일 또는 비밀번호가 잘못되었습니다.' },
+        { status: 401 },
+      );
+    } catch (error) {
+      console.error('Error handling request:', error);
+      return HttpResponse.status(500).json({ message: '서버 오류' });
+    }
+  }),
+
+  http.post('/api/businesses/signin', async ({ request }) => {
+    try {
+      const { email, password } = await request.json();
+
+      if (email === 'test@example.com' && password === 'password123') {
+        const headers = new Headers();
+        headers.set('Authorization', 'mocked-jwt-token');
+        headers.set('Refresh', 'mocked-refresh-token');
+
+        return HttpResponse.json(
+          { message: '로그인 성공' },
+          { status: 200, headers },
+        );
+      }
+
+      return HttpResponse.json(
+        { message: '이메일 또는 비밀번호가 잘못되었습니다.' },
+        { status: 401 },
+      );
+    } catch (error) {
+      console.error('Error handling request:', error);
+      return HttpResponse.status(500).json({ message: '서버 오류' });
+    }
   }),
 ];
