--- src/services/decodingInfo.ts
@@ -1,14 +1,0 @@
-// import base64 from 'base-64';
-
-// const DecodingInfo = (jwtToken: string) => {
-//   let payload = jwtToken.substring(
-//     jwtToken.indexOf('.') + 1,
-//     jwtToken.lastIndexOf('.'),
-//   );
-//   let decodingInfo = base64.decode(payload);
-//   let decodingInfoJson = JSON.parse(decodingInfo);
-
-//   return decodingInfoJson;
-// };
-
-// export default DecodingInfo;
